requirements.txt
streamlit==1.39.0
chromadb==0.5.3
sentence-transformers==3.1.1
PyPDF2==3.0.1
python-dotenv==1.0.1
openai==1.47.1
numpy==2.0.2